fix infinite loading on notifications screen
